# Optimize Integer formatting/parsing