@Gankro 2015-07-18
So I know we went through a lot of effort to basically get top-of-the-line industrial float formatting. I imagine int formatting is a *bit* easier, but I'd be remiss to not ask: are there standard "optimal" algorithms that we should be investing in for this?