@arthurprs 2015-07-18
@Gankro what triggered me to do this work was this repository https://github.com/miloyip/itoa-benchmark which is a good summary of what's out there.

What I have in this PR can be seen as a hybrid of the lut and branchlut versions. My version tries to keep code size small and avoid regressions for small numbers. In addition to that it also reduces 64 arithmetic.

Also note that the biggest increase in speed from lut to branchlut/countlut in that repo are due to avoiding the double buffering, which we can't really avoid in rust due to the fmt machinery.