@Gankro 2015-07-18
I'm pretty sure this merits some new tests for interesting corner-cases.