@arthurprs 2015-07-18
@Gankro thank you very much for your review. I cleaned up the code to account for your suggestions and removed the [last bounds checking](https://github.com/arthurprs/rust-optimized-inttostr/commit/3c4c7e072249f1781fd7a445b9e28c6319ccc768).

As for new tests I'm not aware of any border cases other than non-digits characters (added) and overflows (already present).