@arthurprs 2015-07-18
adjusted the nits and parens, travis is green.