@Gankro 2015-07-18
This is looking great! The only outstanding issue is @eefriedman's concern about the char conversion, which I don't have strong feelings about.