@arthurprs 2015-07-18
I can't find @eefriedman comment here, just in my inbox.

to_digit returns None if the char can't be converted to a number of the specified radix, it sounds simple enough to me. If you guys want I can add the wrapper function with some comments though.