@eefriedman 2015-07-19
GitHub collapses comments against old changes; that's probably why you're not finding it.

The problem isn't that to_digit is weird, it's that converting a UTF-8 byte to a Rust `char` is weird.  Might be enough to just expand the "all valid digits are ASCII" comment a bit.