@arthurprs 2015-07-19
Updated comment

```
    // all valid digits are ascii, so we will just iterate over the utf8 bytes
    // and cast them to chars. .to_digit() will safely return None for anything
    // other than a valid ascii digit a the given radix, including the first-byte
    // of multi-byte sequences.
```