@Gankro 2015-07-19
Ehh the build already started. Unless it bounces spuriously (which is not unheard of), just submit a followup pr to remove the annotation.

Such is the danger of weekend PRs... no queue!